@@ -1,6 +1,7 @@
-'use strict';
+import awsEc2 from './aws-ec2';
+import userpass from './userpass';
 
-module.exports = {
-    'aws-ec2': require('./aws-ec2'),
-    userpass: require('./userpass')
+export default {
+  'aws-ec2': awsEc2,
+  userpass,
 };
