@@ -1,140 +1,50 @@
-'use strict';
+import Bluebird from 'bluebird';
+import lt from 'long-timeout';
+import retry from 'retry';
+import _ from 'lodash';
 
-const async = require('async');
-const backends = require('./backends');
-const joi = require('joi');
-const lt = require('long-timeout');
-const retry  = require('retry');
-const _ = require('lodash');
+import backends from './backends';
 
-module.exports = {
-    /**
-     * Authenticate with vault.
-     * @param  {Object} options
-     * @param  {String} options.backend - ['userpass']
-     * @param  {Function} [cb]
-     */
-    login(options, cb) {
-        const self = this;
-        const _async = _.isFunction(cb);
+export function login(options) {
+  const prior = this.store.get('_auth') || {};
+  const opts = options || prior.login;
 
-        // if this is a renewal, grab prior login options
-        if (!options) {
-            options = _.get(self, '_auth.login');
-        }
+  if (prior.renewal) {
+    lt.clearTimeout(prior.renewal);
+  }
 
-        // clear any pending renewals
-        if (_.has(self, '_auth.renewal')) {
-            lt.clearTimeout(self._auth.renewal);
-            delete self._auth.renewal;
-        }
-
-        async.auto({
-            // validate options
-            options: function validateOptions(fn) {
-                const schema = joi.object({
-                    backend: joi.string().valid([
-                        'aws-ec2',
-                        'userpass'
-                    ]).required(),
-                    options: joi.object().required(),
-                    retry: joi.object()
-                }).unknown(true).required();
-                joi.validate(options, schema, function(err, validated) {
-                    if (err) {
-                        err = new Error(`Invalid Login Options: ${err.messsage}`);
-                        return fn(err);
-                    }
-                    _.set(self, '_auth.login', options);
-                    fn(null, validated);
-                });
-            },
-
-            // attempt to authenticate with vault server, retry with
-            // exponential backoff in case of non authentication errors
-            data: ['options', function authenticate(fn, r) {
-                const backend = backends[options.backend];
-                const ops = r.options;
-                const operation = retry.operation(ops.retry);
-
-                operation.attempt(function() {
-                    backend.login(self, ops.options, function(err, auth) {
-                        if (err) {
-                            const errCode = _.get(err, 'message');
-                            const defaultStatus = /timeout/.test(errCode) ? 503 : 499;
-                            const status = _.get(err, 'status', defaultStatus);
-                            if (status >= 400 && status <= 499) {
-                                if (operation._timeout) {
-                                    clearTimeout(operation._timeout);
-                                }
-                                operation._timeouts = [];
-                                if (status === 400 || status === 401) {
-                                    err = new Error(`Incorrect username/password combination`);
-                                    err.type = 'ERROR:LOGIN';
-                                }
-                                return fn(err);
-                            }
-                            if (operation.retry(err)) {
-                                return;
-                            }
-                            err = operation.mainError() || err;
-                            return fn(err);
-                        }
-
-                        fn(null, auth);
-                    });
-                });
-            }],
-
-            auth: ['data', function validateAuth(fn, r) {
-                const data = r.data;
-                joi.validate(data, joi.object({
-                    client_token: joi.string().required(),
-                    lease_duration: joi.number().integer().min(0).required()
-                }).unknown(true).required(), fn);
-            }],
-
-            // finish login process
-            login: ['auth', function finishLogin(fn, r) {
-                const auth = r.auth;
-
-                self.status = 'authenticated';
-                _.set(self, '_auth.token', auth.client_token);
-                if (auth.lease_duration > 0) {
-                    self._auth.renewal = lt.setTimeout(function() {
-                        self.status = 'unauthenticated';
-                        self.login.call(self);
-                    }, auth.lease_duration * 1000);
-                }
-                async.setImmediate(fn);
-            }]
-        }, function(err) {
-            if (err) {
-                self.emit('error:login', err);
-                if (_async) {
-                    cb(err);
-                }
-                return;
-            }
-            self.emit('authenticated');
-            if (_async) {
-                cb();
-            }
+  return this.validate('login-options', opts)
+  .then(() => this.store.set('_auth.login', opts))
+  .then(() => {
+    return new Bluebird((resolve, reject) => {
+      const backend = backends[options.backend];
+      const operation = retry.operation(options.retry);
+      operation.attempt(() => {
+        backend.login(this, options.options)
+        .then(auth => resolve(auth))
+        .catch((err) => {
+          if (_.get(err, 'response.status', 500) < 500) {
+            operation.stop();
+          }
+          if (operation.retry(err)) {
+            return;
+          }
+          return reject(err);
         });
-    },
+      });
+    })
+    .then(auth => this.validate('auth-response-data', auth))
+    .then((auth) => {
+      this.store.set('_auth.data', auth);
+      if (auth.lease_duration > 0) {
+        this.store.set('_auth.renewal', lt.setTimeout(() => {
+          this.login();
+        }, auth.lease_duration * 1000));
+      }
+    });
+  });
+}
 
+export function logout() {
 
-    /**
-     * Discard all relevant authentication info
-     * @return {[type]}
-     */
-    logout() {
-        const self = this;
-        _.unset(self, '_auth.token');
-        // clear any pending renewals
-        if (_.has(self, '_auth.renewal')) {
-            clearInterval(self._auth.renewal);
-            delete self._auth.renewal;
-        }
-    }
-};
+}
